remove dead code, reorg dedup and sanity
