+++ cloud/bq/dedup.go
@@ -1,4 +1,4 @@
-package dispatch
+package bq
 
 import (
 	"context"
@@ -30,6 +30,7 @@ var (
 )
 
 // WaitForStableTable loops checking until table exists and has no streaming buffer.
+// TODO - move these functions to go/bqext package
 func WaitForStableTable(tt *bigquery.Table) error {
 	errorTimeout := 2 * time.Minute
 	if testMode {
@@ -132,6 +133,7 @@ var dedupTemplateSidestream = `
 //
 // NOTE: If destination table is partitioned, destTable MUST include the partition
 // suffix to avoid accidentally overwriting the entire table.
+// TODO - move these functions to go/bqext package
 func Dedup(dsExt *bqext.Dataset, src string, destTable *bigquery.Table) (*bigquery.Job, error) {
 	if !strings.Contains(destTable.TableID, "$") {
 		meta, err := destTable.Metadata(context.Background())
