+++ cloud/bq/sanity.go
@@ -1,11 +1,11 @@
-// Package dedup provides facilities for deduplicating
-// template tables and copying into a destination partitions.
+// Package sanity provides facilities for sanity checking and
+// copying final table into a destination partition.
 // It is currently somewhat NDT specific:
 //  1. It expects tables to have task_filename field.
 //  2. It expects destination table to be partitioned.
 //  3. It does not explicitly check for schema compatibility,
 //      though it will fail if they are incompatible.
-package dedup
+package bq
 
 import (
 	"errors"
@@ -398,215 +398,4 @@ func SanityCheckAndCopy(ctx context.Context, client *bigquery.Client, srcTable, 
 	err = WaitForJob(context.Background(), job, 10*time.Second)
 	log.Println("Done")
 	return err
-}
-
-// Options provides processing options for Dedup_Alpha
-type Options struct {
-	MinSrcAge     time.Duration // Minimum time since last source modification
-	IgnoreDestAge bool          // Don't check age of destination partition.
-	DryRun        bool          // Do all checks, but don't dedup or copy.
-	CopyOnly      bool          // Skip the dedup step and copy from intermediate to destination
-}
-
-// Job keeps track of a number of resources involved in dedupping.
-type Job struct {
-	// Initialization objects
-	*bqext.Dataset                 // Dataset extension associated with source
-	srcTable       *AnnotatedTable // Source table to be deduplicated
-	destTable      *AnnotatedTable // destination table that result will be committed to.
-
-	// Additional information generated during processing.
-	dedupTable *AnnotatedTable // dedup table that will receive deduped result before commit
-}
-
-// NewJob creates a Job struct
-func NewJob(ds *bqext.Dataset, srcTable, destTable *AnnotatedTable) *Job {
-	return &Job{Dataset: ds, srcTable: srcTable, destTable: destTable}
-}
-
-// This should be called after doing various sanity checks.
-func (job *Job) dedupAndCopy(ctx context.Context, options Options) error {
-	if !options.CopyOnly {
-		// Do the deduplication to intermediate "dedup" table with same root name.
-		// TODO - are we checking for source newer than intermediate destination?  Should we?
-		_, err := job.Dedup_Alpha(job.srcTable.TableID, "test_id", job.dedupTable.Table)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-	}
-
-	// Now compare number of rows and tasks in dedup table to destination table.
-	err := job.dedupTable.checkAlmostAsBig(ctx, job.destTable)
-	if err != nil {
-		return err
-	}
-
-	err = SanityCheckAndCopy(ctx, job.Dataset.BqClient, job.dedupTable.Table, job.destTable.Table)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// CheckAndDedup checks various criteria, and if they all pass, dedups the table.
-// The destination table must exist, must be in a separate dataset, and may or may not
-// have a corresponding partition.  If the partition already exists, then additional
-// sanity checks occur.
-//
-// First checks initial criteria and deduplicates into the appropriate partitioned
-// table corresponding to source.
-// Then if all criteria pass, copies into the destination table.
-//
-// Returns nil if successful, error if criteria fail, or dedup fails.
-//
-// Criteria:
-//   1. Source table modification time is at least as old as options.MinSrcAge
-//   2. Source table mod time is later than destination table mod time, unless options.IgnoreDestAge.
-//   3. Source and destination are in different datasets, and source dataset is NOT "base_tables"
-//   4. If destination partition exists then
-//       4a. Source reflects at least 99% as many task files as final destination
-//       4b. Source has at least 95% as many tests as destination (including dups)
-//       4c. After deduplication, intermediate table has at least 95% as many tests as destination.
-//
-// options - dedup options, MinSrcAge, IgnoreDestAge, DryRun, CopyOnly.
-//
-func (job *Job) CheckAndDedup(ctx context.Context, options Options) error {
-	// Check that source and destination are both templated or partitioned.
-	srcParts, err := getTableParts(job.srcTable.TableID)
-	if err != nil {
-		return err
-	}
-	destParts, err := getTableParts(job.destTable.TableID)
-	if err != nil {
-		return err
-	}
-	// Check that date suffixes match.
-	if destParts.yyyymmdd != srcParts.yyyymmdd {
-		return errors.New("Source and Destination should have same partition/template date: ")
-	}
-
-	// Check if the last update was at least minSrcAge in the past.
-	if time.Since(job.srcTable.LastModifiedTime(ctx)) < options.MinSrcAge {
-		return errors.New("Source is too recent")
-	}
-
-	if job.destTable.DatasetID == job.Dataset.DatasetID {
-		return errors.New("Source and Destination should be in different datasets: ")
-	}
-
-	// For the dedup destination, we use the partitioned table with the same dataset/prefix
-	// as the source table.
-	// We do the deduplication into the partition with the same date.
-	// dedupTable will later be used to create a copier, so it must use a valid bigquery.Client.
-	tmpTable, err := getTable(job.BqClient, job.ProjectID, job.DatasetID, srcParts.prefix, srcParts.yyyymmdd)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-	job.dedupTable = NewAnnotatedTable(tmpTable, job.Dataset)
-
-	// Just check that table exists.
-	_, err = job.srcTable.CachedMeta(ctx)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-
-	// Check that temporary dedup target is neither final destination, nor "base_tables" dataset.
-	if job.dedupTable.DatasetID == job.destTable.DatasetID || job.dedupTable.DatasetID == "base_tables" {
-		return errors.New("Dedup and Destination should be in different datasets: " + job.dedupTable.FullyQualifiedName())
-	}
-
-	if !options.IgnoreDestAge {
-		err = job.srcTable.checkModifiedAfter(ctx, job.destTable)
-		if err != nil {
-			log.Println(err)
-			return err
-		}
-	}
-
-	err = job.srcTable.checkAlmostAsBig(ctx, job.destTable)
-	if err != nil {
-		log.Println(err)
-		return err
-	}
-
-	if options.DryRun {
-		log.Println("Dedup dry run:", job.srcTable.TableID, "test_id", job.dedupTable)
-		return nil
-	}
-
-	err = job.dedupAndCopy(ctx, options)
-
-	// TODO Update status table
-	// We should have a status table that has a row for each table dedup operation.
-	// It should record:
-	//    date, dedup version, requester, cmd params
-	//    source {table_date, mod date, task count, row count},
-	//    precheck outcome
-	//    dedup stats {elapsed time, bytes, rows, error}
-	//    intermediate {table$date, task count, row count, byte count}
-	//    destination {table$date, prev mod date, prev task count, prev row count, prev byte count}
-	//    copycheck outcome
-	//    final copy stats {elapsed time, bytes, rows, error}
-	//    outcome (succeed, precheck failed, dedup failed, copycheck failed, copy failed)
-
-	// TODO If DeleteAfterDedup, then delete the source table.
-	// bq rm 'intermediate$20160301' ??
-	// bq rm 'batch.ndt_20160301'
-
-	return nil
-}
-
-// ProcessTablesMatching lists all tables matching a template pattern, and for
-// any that are at least the age specified in options, dedups and copies them to
-// corresponding partitions in the destination table.
-func ProcessTablesMatching(srcDS *bqext.Dataset, srcPattern string, destDataset string, options Options) error {
-	// This may not have full suffix, so we can't use getTableParts.
-	srcParts := strings.Split(srcPattern, "_")
-	if len(srcParts) != 2 {
-		return errors.New("Invalid source pattern: " + srcPattern)
-	}
-
-	// These are sorted by LastModification, oldest first.
-	atList, err := GetTablesMatching(context.Background(), srcDS, srcPattern)
-	if err != nil {
-		return err
-	}
-
-	// Print a summary of tables to be processed.
-	log.Println("Examining", len(atList), "source tables")
-	for i := range atList {
-		log.Println(atList[i].Table)
-	}
-
-	// Process each table, serially, to avoid problems with too many concurrent
-	// bigquery queries.
-	for i := range atList {
-		srcTable := atList[i]
-		// Skip any partition that has been updated in the past minAge.
-		if time.Since(srcTable.LastModifiedTime(context.Background())) < options.MinSrcAge {
-			continue
-		}
-
-		srcParts, err := getTableParts(srcTable.TableID)
-		if err != nil {
-			return err
-		}
-
-		// destination should be in a different dataset, but same project and same table name and
-		// date suffix as source.
-		destTable, _ := getTable(srcDS.BqClient, srcDS.ProjectID,
-			destDataset, srcParts.prefix, srcParts.yyyymmdd)
-		job := NewJob(srcDS, &srcTable, NewAnnotatedTable(destTable, srcDS))
-		err = job.CheckAndDedup(context.Background(), options)
-		if err != nil {
-			log.Println(err, "dedupping", srcDS.DatasetID+"."+srcTable.TableID,
-				"to", destTable.DatasetID+"."+destTable.TableID)
-			return err
-		}
-	}
-	return nil
 }
